requirements.txt
aerich==0.7.2
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
atlastk==0.13.3
click==8.1.7
dictdiffer==0.9.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
loguru==0.7.2
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
pypika-tortoise==0.1.6
python-multipart==0.0.9
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
toml==0.10.2
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.29.0
